# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NUS API Documentation",
        "description": "NUS Swagger OpenApi description",
        "contact": {
            "email": "ryanhaidypapudi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get list of questions",
                "description": "",
                "operationId": "GetQuestions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Store question to system",
                "description": "",
                "operationId": "PostQuestion",
                "requestBody": {
                    "description": "Question store request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionStoreRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/questions/{slug}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get question by slug",
                "description": "",
                "operationId": "GetQuestion",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of question to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update question in system",
                "description": "",
                "operationId": "PutQuestion",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of question to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Question update request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionUpdateRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Question"
                ],
                "summary": "Delete question in system",
                "description": "",
                "operationId": "DeleteQuestion",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of question to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/questions/follow": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Add/Remove user's followed question in system",
                "description": "",
                "operationId": "FollowQuestion",
                "requestBody": {
                    "description": "Question follow request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionFollowRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionFollowRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuestionFollowRequest": {
                "properties": {
                    "question_id": {
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "QuestionStoreRequest": {
                "properties": {
                    "topic_id": {
                        "type": "int"
                    },
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "QuestionUpdateRequest": {
                "properties": {
                    "topic_id": {
                        "type": "int"
                    },
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}